requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djoser==2.3.3
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
social-auth-app-django==5.7.0
Pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.9.0
django-storages==1.14.4
boto3==1.35.90
drf-yasg==1.21.8

# Production dependencies
sentry-sdk[django]==2.19.2
